% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mice.impute.truescore}
\alias{mice.impute.truescore}
\title{True Score Imputation with mice}
\usage{
mice.impute.truescore(y, ry, x, wy = NULL, calibration = NULL, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{calibration}{A list of calibration information used for true score
imputation. See below for details.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
This custom imputation function is used with the \code{mice}
package by setting method='truescore' for each variable imputed using
true score imputation, which will call this custom imputation function
\code{mice.impute.truescore}. Although possible, this function is not
meant to be run on its own; see documentation for other \code{mice}
imputation files, e.g., \code{\link[mice]{mice.impute.pmm}}, for details
on this usage. Example usage through the mice package is provided in
Examples below.
}
\section{Passing Calibration Information to \code{mice}}{


The \code{calibration} parameter is passed to the \code{mice} function
using the \code{blots} input. For each imputed true score, provide the
\code{calibration} information as a named list. The following elements
are required, in any order:

\describe{
  \item{\code{OSNAME}}{Name of the variable in the data set containing the
  observed scores used for true score imputation}
  \item{\code{scoreType}}{Type of score provided. Current options are
  \code{'CTT'}, corresponding to the classical test theory model of
  reliability; \code{'EAP'}, corresponding to expected a posteriori
  scoring in item response theory; and \code{'ML'}, corresponding to
  maximum likelihood scoring in item response theory. Each
  \code{scoreType} requires specific other elements to be provided
  in \code{calibration} data; see below for these conditional elements.}
  \item{\code{mean}}{The mean of the score metric from calibration. For example,
  T scores are calibrated to a mean of 50, so if T scores are used,
  \code{mean} should be set to \code{50}.}
  \item{\code{varTS}}{The variance of the score metric from calibration. For
  example, T scores are calibrated to a standard deviation of 10, so
  if T scores are used, \code{varTS} should be set to \code{100}.}
}

In addition, each \code{scoreType} requires specific other elements
to be provided in \code{calibration} data:

\describe{
  \item{\code{SENAME}}{Required if \code{scoreType == 'EAP'} or
  \code{scoreType == 'ML'}. Name of the variable in the data set
  containing the standard error estimates of the observed scores
  provided in \code{OSNAME}.}
  \item{\code{reliability}}{Required if \code{scoreType == 'CTT'}.
  Reliability estimate denoting the ratio of true score to
  observed score variance, as estimated from calibration.}
}
}

\section{Specifying the Predictor Matrix}{

I dunno lol
}

\examples{
hihi hello
}
